6-7 done

+++ Domain/GoodsDomain/GoodsModel.cs
@@ -4,11 +4,11 @@ using Types;
 namespace Domain.GoodsDomain
 {
     public abstract class GoodsModel
-    {
-        public int Id { get; set; }
-        public int Price { get; set; }
+    { 
+	    public int Id { get; set; }
+        public double Price { get; set; }
         public string Name { get; set; }
-
+		
         public abstract GoodsType Type { get; }
         public abstract TimeSpan ProcessTime { get; }
 
@@ -18,5 +18,6 @@ namespace Domain.GoodsDomain
                    " Price: " + Price + " | " +
                    " Type: " + Type;
         }
+        
     }
 }
